Rename seedadmission component to seedadmissioncontroller
